# Add optimized lock methods for `Sharded` and refactor `Lock`